fix: resolve compiler warnings in clip application
- button.c: use %ld for atomic_val_t format specifier
- audio.c: cast uint64_t to unsigned int for %u format
- config.h: include clip.h for enum recording_mode definition
- at_commands.c: remove unused cmd_noise_handler and cmd_dereverb_handler

+++ applications/clip/src/audio.c
@@ -894,8 +894,8 @@ create_new_segment:
 
         LOG_INF("rec done: %u frm %us %uKB",
             stats.frames_encoded,
-            stats.recording_time_ms / 1000,
-            stats.total_bytes / 1024);
+            (unsigned int)(stats.recording_time_ms / 1000),
+            (unsigned int)(stats.total_bytes / 1024));
     }
 }
